% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_faf_to_zones.R
\name{allocate_faf_to_zones}
\alias{allocate_faf_to_zones}
\title{Allocate FAF daily truck flows to traffic analysis zones}
\usage{
allocate_faf_to_zones(
  daily_faf_trips,
  synthetic_firms,
  makeuse,
  intermodal_connectors,
  swim_external_zones = 5001:5012
)
}
\arguments{
\item{daily_faf_trips}{Data frame containing discrete daily FAF truck trips}

\item{synthetic_firms}{Data frame containing the synthetic firms and their
attributes, either exogenously specified or created within this platform}

\item{makeuse}{Data frame containing input-output make and use coefficient,
previously translated from PECAS format to that usable by this platform}

\item{intermodal_connectors}{Data frame containing the location (traffic
analysis zone) of air and sea ports within the modeled airport, and
percentage of such flows originating and terminating within the FAF region
they are situated in should be allocated to them}

\item{external_gateways}{Data frame containing the external stations (traffic
analysis zone) that truck flows to or from FAF regions outside of the
modeled area enter or leave through}
}
\description{
Allocate FAF daily truck flows to traffic analysis zones
}
\details{
This function allocates discrete daily FAF (inter-regional) truck
  flows, coded to FAF regions, to traffic analysis zones within the modeled
  area. Input-output coefficients are used to map the commodity carried by
  each truck (or formerly or next carried, in case of empty trucks) to the
  industry type most likely to produce or consume it. Employment within the
  chosen industry within all zones situated within the FAF region are sampled
  to choose the internal trip end. Finally, the flows are assigned to
  external gateways entering or leaving the region. For trucks, external FAF
  regions are associated with a particular gateway. Note that through truck
  trips are coded to two external gateways. For import or export flows the
  appropriate intermodal connector (air or sea port) is chosen instead of
  internal zone within the modeled area. The result is a complete truck trip
  record at the traffic analysis zone level, for each daily truck trip in the
  dataset.
}
\examples{
daily_faf_flows <- allocate_faf_to_zones(daily_faf_trips, synthetic_firms,
  makeuse, intermodal_connectors, external_gateways)
}
