% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_weighted_distances.R
\name{append_weighted_distances}
\alias{append_weighted_distances}
\title{Added weighted distances to FAF 5.x+ data without it}
\usage{
append_weighted_distances(
  fhwa_db,
  distances,
  internal_regions = c(411, 419),
  ignore_regions = c(151, 159)
)
}
\arguments{
\item{fhwa_db}{A FAF regional database}

\item{distances}{A tibble containing intra-regional weighted distances
(`wgt_dist`) by `dms_orig` and `dms_dest` pairs}

\item{internal_regions}{A list of FAF regions within the SWIM2 modeled area
(defaults to FAF regions 411 [Portland] and 419 [rest of Oregon])}

\item{ignore_regions}{A list of FAF regions that will be ignored if weighted
distances are not found in the data (defaults to currently defined FAF 
regions in Hawaii, which are not reachable by surface transport modes)}
}
\description{
Added weighted distances to FAF 5.x+ data without it
}
\details{
This appends weighted distances from an exogenously produced file of
  FAF region interchanges. The `distances` tibble or file is assumed to 
  contain only those fields (i.e., `dms_orig`, `dms_dest`, `wgt_dist`). This
  tibble is joined to the FAF regional database records, and the resulting
  tibble (the original FAF regional database plus the new `wgt_dist` field)
  saved for later use. If an interchange is present in `fhwa_db` that is not
  found in `distances` the program crashes. In that case the user will need to
  add the missing interchanges to `distances` before continuing.
}
\examples{
add_distances <- append_weighted_distances("./faf5.3.csv.xz",
  "./faf451_distances.csv")
}
