% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annual_truckload_equivalents.R
\name{create_annual_truckload_equivalents}
\alias{create_annual_truckload_equivalents}
\title{Generate annual truckload equivalents using CVS and FAF4 truck factors}
\usage{
create_annual_truckload_equivalents(
  faf_flows,
  truck_allocation_factors,
  payload_distributions,
  stop_frequencies = NULL,
  cache_size = 1e+06,
  trace_flows = FALSE
)
}
\arguments{
\item{faf_flows}{A tibble containing FAF truck commodity flow records for
the current model year}

\item{truck_allocation_factors}{A tibble containing factors that map
commodity flows to different truck types}

\item{payload_distributions}{A tibble containing the frequency or percentage
of trucks by payload weight in the population for each distance band in the
data}

\item{stop_frequencies}{A tibble containing the incidence of number of stops
per truck tour (OD movement in the FAF data) by distance band (optional,
defaults to NULL indicating that no stop frequencies will be applied)}

\item{cache_size}{An integer specifying the number of sampled trucks to be
generated within each distance band (defaults to 1E6)}

\item{trace_flows}{Turns detailed tracing statements on or off (defaults to
FALSE)}
}
\description{
Generate annual truckload equivalents using CVS and FAF4 truck factors
}
\details{
This function creates a database of discrete truckload equivalents
  corresponding to each FAF commodity flow record. This conversion is meant
  to be carried out for trucks only, but could be expanded to include mixed
  modes of shipment that include trucking. This process represents a
  departure from earlier methods based on the FAF4 Freight Traffic
  Assignment report. Instead, trucks by type are generated using the FAF
  truck allocation factors with cargo weights sampled from observed distri-
  butions from the 2012 Canadian Commercial Vehicle Survey (CVS). A data-
  base of discrete trucks are created for each FAF commodity flow, with
  the sum of tonnage and value for corresponding synthetic trucks equal to
  the same values reported in the FAF.
}
\examples{
faf_annual_trucks <- create_annual_truckload_equivalents(faf_data,
  truck_allocation_factors, cvs_payload_weight_distributions,
  cvs_stop_pattern_frequencies)
}
