% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synthetic_firms.R
\name{create_synthetic_firms}
\alias{create_synthetic_firms}
\title{Create synthetic pseudo-firms}
\usage{
create_synthetic_firms(
  zonal_data,
  sector_equivalencies,
  alpha2beta,
  save_to = NULL
)
}
\arguments{
\item{zonal_data}{Data frame containing traffic analysis zone attributes, to
include employment by sector}

\item{sector_equivalencies}{Data frame containing PECAS sectors and their
corresponding firm categories (groups of sectors) used in trip generation}

\item{alpha2beta}{Data frame containing, among other things, the county and
Freight Analysis Framework (FAF) region that each zone lies within}

\item{save_to}{File name for saving the pseudo-firms produced by this 
function (optional)}
}
\description{
Create synthetic pseudo-firms
}
\details{
This function maps employment by PECAS sector within each alpha zone
  into separate pseudo-firms, which are unit of analyses for remainder of CT
  simulation.
}
\examples{
zones <- create_synthetic_firms(zonal_data, sector_equivalencies, alpha2beta,
  "synthetic-firms.csv")
}
