% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_helpers.R
\name{ct_msg}
\alias{ct_msg}
\title{Print a ct message}
\usage{
ct_msg(s = NULL, header = NULL)
}
\arguments{
\item{s}{Character string to print to console}

\item{header}{[Optional] string to print as a header

Instead of R's normal print function we can use one that prints without quotes,
as well as extending it to print function headers in the output.}
}
\description{
Print a ct message
}
