% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_trip_list.R
\name{export_trip_list}
\alias{export_trip_list}
\title{Export local and inter-regional trip records in a combined trip list file}
\usage{
export_trip_list(hourly_faf_trips, hourly_local_trips, save_to)
}
\arguments{
\item{hourly_faf_trips}{Data frame containing the final inter-regional flows,
with origin, destination, and departure time appended for each trip}

\item{hourly_local_trips}{Data frame containing the final local (internal)'
truck flows, with origin, destination, and departure time appended for each
trip}

\item{save_to}{Filename for saving the final combined trip list in comma-
separated value (CSV) format}
}
\description{
Export local and inter-regional trip records in a combined trip list file
}
\details{
This function will typically be the last part of the model to run,
  for it combines the local and inter-regional truck trip records into a
  single trip file, with common fields for both. Missing values are inserted
  for data present in one dataset, but not in the other. The result is saved
  in a file for network analyses or post-processing. No values are returned
  by the function.
}
\examples{
export_trip_list(hourly_faf_trips, hourly_local_trips, "Trips_CTTruck.csv")
}

