% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tibble_contents.R
\name{list_tibble_contents}
\alias{list_tibble_contents}
\title{List the class and range of values in each column of a tibble}
\usage{
list_tibble_contents(df, threshold = 10)
}
\arguments{
\item{df}{Tibble (data frame) containing the variables to be summarized}

\item{threshold}{The number of discrete values of a certain variable to
enumerate. Defaults to 10, which typically fits within margin of knitr
output}
}
\description{
List the class and range of values in each column of a tibble
}
\details{
This function creates a tibble with the name, class, and range
  of values included for each column, as well as the number of missing and
  zero values. It returns a tibble with that information, with one variable
  per record.
}
\examples{
df_contents <- list_tibble_contents(my_tibble, 9)
}
