% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annual_faf_data.R
\name{load_annual_faf_data}
\alias{load_annual_faf_data}
\title{Load preformatted and processed FAF data for the target year}
\usage{
load_annual_faf_data(preprocessed_faf, target_year, interpolate = FALSE)
}
\arguments{
\item{preprocessed_faf}{Data frame containing the FAF inter-regional flow
database previous extracted for modeled area and in format required by CT
modules}

\item{target_year}{The four-digit simulation year as an integer value}

\item{interpolate}{A boolean variable denoting whether values for the
`target_year` should be interpolated between the two closest years if it is
not one of the years included in the FAF database (defaults to FALSE,
currently ignored)}
}
\description{
Load preformatted and processed FAF data for the target year
}
\details{
This function retrieves FAF regional flow data for a specific year
  from data already preprocessed from the original format distributed by
  FHWA. This conversion is usually done only once per FAF version and then
  stored as static inputs to the SWIM2 system. These preprocessed data are
  typically built using the `prebuild_faf_multiyear.rmd` program in the
  `data-raw` folder of the `swimctr` package. When used in a CT run within
  this module obviates the need to run the `swimctr::preprocess_faf_database`
  module, saving consider CT run time for each simulation year.
}
\examples{
annual_flows <- load_annual_faf_data(prebuilt_database, 2018)
}
