% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tabular_data.R
\name{load_tabular_data}
\alias{load_tabular_data}
\title{Read tabular data from an ambiguous data source}
\usage{
load_tabular_data(data_object, sep = ",", comment = "#", verbose = FALSE)
}
\arguments{
\item{data_object}{A fully-qualified path for a file or a tibble previously
created and visible within the current environment}

\item{sep}{A character that delimits the values in an external file. Only
applicable if a filename is passed to the function, and defaults to a
comma.}

\item{comment}{A leading character that identifies the information on a
 record as a comment. Only applicable if a filename is passed to the
function, and defaults to '#'.}

\item{verbose}{Prints a listing of the variables, class, number of unique
values, and range of values for the tibble if TRUE. Defaults to FALSE.}
}
\description{
Read tabular data from an ambiguous data source
}
\details{
This function reads, and optionally lists a summary of, tabular data
  from an ambiguous source. That is, the data object can either be passed in
  a tibble or string name. If the latter the file's contents are read into
  the tibble and returned to the calling program. Thus, the user can pass
  either without having to change the code the calls this function. Moreover,
  different delimiters can be used without having to convert the source data
  into CSV (e.g., semi-colon or tab).
}
\examples{
this_tibble <- load_tabular_data(another_tibble)  # Does nothing
this_tibble <- load_tabular_data("./myfile.csv.xz", verbose = TRUE)
}
