% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_truck_generation.R
\name{local_truck_generation}
\alias{local_truck_generation}
\title{Generate local truck trips (tour segments) from pseudo-firms}
\usage{
local_truck_generation(
  synthetic_firms,
  generation_probabilities,
  random_seed = 1,
  max_resampling_attempts = 30,
  max_resampling_threshold = 0.5
)
}
\arguments{
\item{synthetic_firms}{Data frame containing pseudo-firms and their
typical attributes}

\item{generation_probabilities}{Data frame containing trip generation rates
expressed as probabilities, for each trip generation category used in the
model}

\item{random_seed}{Integer value of random number seed set at outset of
sampling of trips generated by pseudo-firms (defaults to 1)}

\item{max_resampling_attempts}{Integer value that controls the number of
times the generation sampling will repeat if it does not finish with same
number of trips calculated in aggregate for the entire study area. If this
number is exceeded then the model will error. Alternatively if `1` then
resampling will be turned off.}

\item{max_resampling_threshold}{Floating point number defining how close
the simulation must match aggregate number of trips for the entire
modeled area (default is 0.5 percent)}
}
\description{
Generate local truck trips (tour segments) from pseudo-firms
}
\details{
The trip generation function samples the number of discrete daily
  trips generated by each pseudo-firm, and appends that number to each
  synthetic firm. The mode of transport used for each trip is also sampled
  and added to the trip records.
}
\examples{
daily_trips <- local_truck_generation(synthetic_firms,
  generation_probabilities)
daily_trips <- local_truck_generation(synthetic_firms,
  generation_probabilities, random_seed = 1, max_resampling_attempts = 30,
  max_resampling_threshold = 0.5)
}
