% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_faf4_database.R
\name{preprocess_faf4_database}
\alias{preprocess_faf4_database}
\title{Transform FAF flow database distributed by FHWA into format for target year}
\usage{
preprocess_faf4_database(fhwa_db, target_year)
}
\arguments{
\item{fhwa_db}{Data frame containing the FAF interregional flow database in
format distributed by FHWA}

\item{target_year}{The four-digit simulation year as an integer value. The
value must be one of the years in the FAF database, or fall within the
interval x years before first year to x years after the time series in the
FAF data, where x is the foreast_range_threshold.}
}
\description{
Transform FAF flow database distributed by FHWA into format for target year
}
\details{
This function converts the commodity flows, measured in annual tons
  and value between FAF regions, into flows for the specified year or closest
  to it in the FAF database.
}
\examples{
annual_flows <- preprocess_faf4_database(fhwa_database, 2018)
}
