% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_external_faf_regions.R
\name{recode_external_faf_regions}
\alias{recode_external_faf_regions}
\title{Recode FAF regions beyond the SWIM halo to SWIM external zones}
\usage{
recode_external_faf_regions(
  annual_faf_flows,
  external_equivalencies,
  through_equivalencies,
  swim_faf_regions = c(411, 419, 532)
)
}
\arguments{
\item{annual_faf_flows}{A tibble or name of file containing pre-
processed annual FAF commodity flows by trucks for the desired year}

\item{external_equivalencies}{A tibble or name of file containing
equivalencies between distal FAF regions and SWIM external stations}

\item{through_equivalencies}{A tibble or name of file containing
equivalencies between distal FAF regions for both trip ends and the likely
SWIM external zones they pass through}

\item{swim_faf_regions}{A list of FAF regions that are wholly contained
within the SWIM modeled area}
}
\description{
Recode FAF regions beyond the SWIM halo to SWIM external zones
}
\details{
This function replaces the domestic origin and destination
  (`dms_orig` and `dms_dest`) of FAF regions outside the SWIM modeled area
  with their corresponding SWIM external stations. In most cases two or more
  candidate routes will be available, each with an exogenously defined
  probability of being chosen based on user-supplied travel times between the
  centroids of the FAF regions. This function will proportionally assign
  commodity flows, in annual tons, between the FAF regions. Note that FAF
  reqions within the modeled area, defined in `swim_faf_regions` are not
  changed. They will be allocated to alpha zones later in the CT process
  after these flows have been transformed into annual truckload equivalents.

  The `swim_faf_regions` correspond to FAF regions that are completely
  contained within the SWIM modeled area. This is currently defined as the
  Oregon counties within the Portland MSA (FAF region 411), the remaining
  counties in Oregon (FAF region 419), and the Washington counties within the
  Portland MSA (FAF region 532).
}
\examples{
recode_to_externals <- recode_external_faf_regions(annual_faf_flows, 2025,
  external_equivalencies, through_equivalencies)
}
