% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_daily_faf_flows.R
\name{sample_daily_faf_flows}
\alias{sample_daily_faf_flows}
\title{Sample daily FAF flows from annual truckload equivalents}
\usage{
sample_daily_faf_flows(
  annual_faf_flows,
  target_year,
  external_targets,
  internal_faf_regions = c(411, 419, 532),
  daily_scaling_factor = 0.001274154
)
}
\arguments{
\item{annual_faf_flows}{Tibble containing annual FAF truckload equivalents}

\item{target_year}{Integer value of the current SWIM2 simulation year}

\item{external_targets}{Tibble containing the target entry and exit truck
volumes at the FAF external zones for each possible `t.year`}

\item{daily_scaling_factor}{Real number the specifies the proportion of FAF
annual truckload trips that will occur on the simulation day, defaults to
calibrated value that best reproduces 2020 observed flows}

\item{internal_faf_region}{A list of the FAF regions included within the SWIM
halo, defaults to 411 (Oregon portion of Portland MSA), 419 (remainder of
Oregon), and 532 (Washington portion of Portland MSA).}
}
\description{
Sample daily FAF flows from annual truckload equivalents
}
\details{
This function samples daily FAF truck trips from the previously
  created annual FAF truckload equivalents. The user will specify the truck
  volumes at each SWIM external station, which will constrain the sampled
  annual flows. A separate user-defined value must be supplied by the user to
  sample the annual internal flows.
}
\examples{
daily_faf_flows <- sample_daily_FAF_flows(annual_faf_flows, external_targets,
  daily_scaling_factor = 0.00385)
}
