% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_daily_faf_truckloads.R
\name{sample_daily_faf_truckloads}
\alias{sample_daily_faf_truckloads}
\title{Sample daily truck trips from annual FAF flow database}
\usage{
sample_daily_faf_truckloads(annual_faf_trucks, target_day = 4)
}
\arguments{
\item{annual_faf_trucks}{A tibble containing the annual truckload equivalents
(discrete truck records) generated from the FAF commodity flow database}

\item{target_day}{The day of the week the flows are sampled for (default = 4,
which corresponds to Wednesday)}
}
\description{
Sample daily truck trips from annual FAF flow database
}
\details{
This function samples daily truck trips from FAF annual truckload
  equivalents. It samples for a designated day of the week in the first week
  of the year, as the current version does not include monthly or weekly flow
  percentages by SCTG or other variables. A tibble is returned with the
  sampled daily truck trips.
}
\examples{
daily_trucks <- sample_daily_faf_truckloads(annual_faf_truckloads, 5)
}
