% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_zonal_households.R
\name{summarize_zonal_households}
\alias{summarize_zonal_households}
\title{Summarize number of households per traffic analysis zone}
\usage{
summarize_zonal_households(zonal_data)
}
\arguments{
\item{zonal_data}{A tibble containing zonal population and employment
estimates at the traffic analysis zone level}
}
\description{
Summarize number of households per traffic analysis zone
}
\details{
This simple function tallies the number of households within each
  traffic analysis zone within the modeled area. These totals are included in
  the calculation of local truck attraction proportions. The result is a data
  frame containing this information, which can optionally be also stored in a
  comma-separated value file if desired.
}
\examples{
total_households <- summarize_zonal_households(zonal_data)
}
