% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annual_faf_truckloads.R
\name{create_annual_faf_truckloads}
\alias{create_annual_faf_truckloads}
\title{Generate annual truckload equivalents using FAF3 Traffic Analysis process}
\usage{
create_annual_faf_truckloads(
  faf_data,
  truck_allocation_factors,
  truck_equivalency_factors,
  empty_truck_factors,
  debug_trace = FALSE
)
}
\arguments{
\item{faf_data}{Data frame containing reformatted FAF commodity flow database}

\item{truck_allocation_factors}{Data frame containing factors that map
commodity flows to different types of trucks}

\item{truck_equivalency_factors}{Data frame containing factors that
translate tons into truckload equivalents}

\item{empty_truck_factors}{Data frame containing factors for calculating
incidence of empty trucks, given the trade (domestic versus foreign) and
truck type}

\item{debug_trace}{Turns detailed tracing statements on or off (defaults to
FALSE)}
}
\description{
Generate annual truckload equivalents using FAF3 Traffic Analysis process
}
\details{
This function appends annual truckload equivalents to the FHWA
  Freight Analysis Framework (FAF) commodity flow forecast database. The
  process is based upon the FAF3 Traffic Analysis report developed by
  Battelle, which maps tonnage to truckload equivalents, including empty
  trucks. The function requires reformatted FAF data that replaces several
  numeric values with strings that are more intuitive and retains data only
  for the target year. The function also requires three tables from Chapter 3
  of the FAF Traffic Analysis report, and then follows their methodology
  exactly (but without the errors in the report). The function returns an
  updated FAF database with truckload equivalents appended to each record.
}
\examples{
faf_annual_trucks <- create_annual_faf_truckloads(faf_data,
  truck_allocation_factors, truck_equivalency_factors, empty_truck_factors)
}
