% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runtime_parameters.R
\name{get_runtime_parameters}
\alias{get_runtime_parameters}
\title{Read the runtime parameters file and store them in environment}
\usage{
get_runtime_parameters(properties_FN, file_format = "simple", delimiter = "=")
}
\arguments{
\item{properties_FN}{Filename containing the runtime parameters, with full
path information included in the filename}

\item{file_format}{The format the runtime parameters are stored in within the
file, which includes 'simple' (the default, and currently only supported
format) and hypertext formats in the future}

\item{delimiter}{The character used to separate the parameter name from its
value (defaults to '=', not relevant for XML format)}
}
\value{
This function creates a RTP environment that is subsequently accessed by
  all CT functions during a model run. The function returns a string that
  reports the tokens and their corresponding values. The user can ignore the
  results by calling the function with reading the results into a variable.
}
\description{
Read the runtime parameters file and store them in environment
}
\details{
This function reads a file containing the runtime parameters
  required by the CT platform. The variables required depend somewhat upon
  the exact model implementation (i.e., what CT functions are called, using
  what methods, and in what order). These should be specified by the model
  developer. Note that it is possible to call this function multiple times in
  order to set default parameters the first time through, and scenario- or
  run-specific parameters later. In that case, earlier parameters are
  overwritten if also defined in files read later. This enables the latter to
  only contain parameter values that differ from one model run to the next,
  rather than repeating standard parameters that rarely change, if ever,
  between different simulations.

  A token name can be included in one or more value fields by enclosing it
  between percent signs. For example, the `t.year` parameters could be
  included in the value `results_%t.year%.csv`. All of the token substitution
  is carried out before the token-value pairs are placed in global runtime
  environment.
}
\examples{
get_runtime_parameters("/Models/swim25/run81/t0/swim.properties")
first_params <- get_runtime_parameters(
  "/Models/swim25/run91/t25/swim.properties")
}
