% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_skim_matrices.R
\name{read_skim_matrices}
\alias{read_skim_matrices}
\title{Convert time and distance skim matrices in OMX format to data frame format}
\usage{
read_skim_matrices(distance_skimsFN = NULL, time_skimsFN = NULL)
}
\arguments{
\item{distance_skimsFN}{Fully-qualified filename of distance skim matrix in
ZMX format}

\item{time_skimsFN}{Fully-qualified filename of travel time skim matrix in
ZMX format}
}
\value{
A tibble containing both time and distance skims on each record.
}
\description{
Convert time and distance skim matrices in OMX format to data frame format
}
\details{
This function reads distance and travel time skim matrices stored
  in compressed matrix format (zmx) used by the SWIM system. The function
  does not assume that the dimensionality is the same, so as to not use
  column joins that might not properly align. Instead, the more inefficient
  but safer method of matching the two by origin and destination zone
  numbers are used, with missing values inserted for when a zone pair
  appears in one file but not the other.
}
\examples{
skim_matrices <- read_skim_matrices("pmautodist.zmx", "pmautotime.zmx")
}
