% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_daily_faf_truckloads.R
\name{sample_daily_faf_truckloads}
\alias{sample_daily_faf_truckloads}
\title{Sample daily FAF truck trips from annual FAF flow database}
\usage{
sample_daily_faf_truckloads(annual_faf_trucks,
  external_scaling_factor = 1, weeks_per_year = 50, target_week = 24,
  target_day = 4, cluster_log = "cluster.log", save_to = NULL)
}
\arguments{
\item{annual_faf_trucks}{Data frame containing reformatted FAF commodity flow
database with estimate of total annual truckload equivalents appended}

\item{external_scaling_factor}{Asserted or estimating factor to account for
FAF commodity flows that grouped into single truck movements (default is
1.0, which assumes no such factoring is carried out}

\item{weeks_per_year}{The number of weeks per year the annual flows are 
assumed to be spread across (default = 50, which is less than full year,
to account for reduced flows on national holidays)}

\item{target_week}{The week of the year that the simulated flows will be
extracted for (default = 24, which is near end of April)}

\item{target_day}{The day of the week the flows are sampled for (default = 4,
which corresponds to Wednesday)}

\item{cluster_log}{The filename of a file used to store output from the
doParallel cluster (optional)}

\item{save_to}{File name for saving the daily FAF truck trip records in
comma-separated value format (optional)}
}
\description{
Sample daily FAF truck trips from annual FAF flow database
}
\details{
This function appends annual truckload equivalents to the FHWA
  Freight Analysis Framework (FAF) commodity flow forecast database. The
  process is based upon the FAF3 Traffic Analysis report developed by 
  Battelle, which maps tonnage to truckload equivalents, including empty
  trucks. The function requires reformatted FAF data that replaces several
  numeric values with strings that are more intuitive, adds the distance
  traveled between FAF region centroids to each record, and converts the data
  from the wide format FHWA distributes it in to tall format, with separate
  record for each foreast year for each origin-destination-mode-commodity
  combination. The function also requires three tables from Chapter 3 of the
  FAF Traffic Analysis report, and then follows their methodology exactly 
  (but without the errors in the report). Because the SWIM simulation runs
  through time the data are mapped to the closest FAF forecast year for each
  year in the simulation. The function returns an updated FAF database with
  truckload equivalents appended to each record.
}
\examples{
faf_annual_trucks <- create_annual_faf_truckloads(faf_data,
  truck_allocation_factors, truck_equivalency_factors, empty_truck_factors,
  2020)  # Accepts the default for cluster log and does not save output
}
